NavigationBar에서 ProfileButton을 별도 파일로 분리하고 불필요한 import 를 제거했습니다. 이제 NavBar가 다시 server component가 되었습니다
